@@ -1,5 +1,3 @@
 module github.com/nginxinc/nginx-plus-go-client
 
-go 1.19
-
-require golang.org/x/exp v0.0.0-20230905200255-921286631fa9
+go 1.21.2
